// src/interfaces/artefact.request.ts
export interface ArtefactRequest {
  id: string;
  name: string;
  description: string;
  type: string;
  level: number;
  subtype: string;
  alias: string;
  category: string;
  subcategory: string;
  version: string;
  company: string;
  owner: string;
  state: string;
  objetive: string;
  range: string;
}